Added PlayerNetworkedData script
New PlayerNetworkedData.cs script is a NetworkBehavior script that should work as an intermediary to sync data between server and clients.

Currently used to hold a copy of playerList that should be synced between server and clients.

Also added several new events for when a player is added or removed.